harp-proxy/pyproject.toml
[tool.poetry]
name = "harp_benchmark"
version = "0.1.0"
description = ""
authors = ["Romain Dorgueil <romain@makersquad.fr>"]
readme = "README.rst"

[tool.poetry.dependencies]
python = "^3.12"
harp-proxy = { path="../../harp", develop=true }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
